Feat center/lostitem/detail 자기글 아니면 삭제버튼 안보이게 수정, center/oneonone/list 로긴안하면 못들가게 수정

+++ src/main/webapp/center/lostitem/detail.jsp
@@ -1,8 +1,22 @@
+<%@page import="java.net.URLEncoder"%>
+<%@page import="vo.Member"%>
+<%@page import="dao.MemberDao"%>
 <%@page import="vo.Lostitem"%>
 <%@page import="dao.LostitemDao"%>
 <%@page import="java.util.List"%>
 <%@ page contentType="text/html; charset=utf-8" pageEncoding="utf-8" %>
 <%
+
+	//세션에서 로그인된 사용자 정보를 조회한다.
+	String id = (String) session.getAttribute("loginId");
+
+	MemberDao memberDao = MemberDao.getInstance();
+	Member member = memberDao.getMemberById(id);
+	
+	if (member == null) {
+		response.sendRedirect("../../member/loginform.jsp?err=req&job="+URLEncoder.encode("문의글 확인", "utf-8"));
+		return;
+	}
 
 	int no = Integer.parseInt(request.getParameter("no"));
 	
@@ -55,7 +69,11 @@
 		
 		<div style="text-align: center; padding:30px;">
 				<a href="list.jsp" class="btn btn-secondary btn-sm">목록</a>
+				
+<% if (lostitem.getMember().getId().equals(id)) {%>				
 				<a href="delete.jsp?no=<%=lostitem.getNo() %>" class="btn btn-secondary btn-sm">삭제</a>
+<% } %>				
+				
 		</div>
